testing/pipeline/requirements.txt
ansible==2.9.5
astroid==2.3.3
atlassian-python-api==1.13.19
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
fabric==2.3.1
google-auth==1.11.2
idna==2.9
invoke==1.4.1
isort==4.3.21
Jinja2==2.11.1
kubernetes==10.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
paramiko==2.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pymongo==3.7.1
PyNaCl==1.3.0
python-dateutil==2.8.1
pyvmomi==6.7.3
PyYAML==5.3
randmac==0.0.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
websocket-client==0.57.0
wrapt==1.11.2
